% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{ccPlots}
\alias{ccPlots}
\title{Plot Faceted cell cycle plots}
\usage{
ccPlots(input, output, session, seu)
}
\arguments{
\item{seu}{Seurat object}

\item{featureType}{Gene or Transcript}

\item{organism_type}{Organism}
}
\description{
Plot Faceted cell cycle plots
}
