% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cc}
\alias{plot_cc}
\title{Plot CC}
\usage{
plot_cc(
  object = NULL,
  group.by = "gene_snn_res.0.6",
  assay = "gene",
  label = "_filtered_",
  color.by = "batch",
  mytitle = "Title",
  faceted = TRUE,
  pt.size = 1,
  phase_levels = c("g1", "g1_s", "s", "s_g2", "g2", "g2_m", "post_mitotic", "hsp",
    "g1_stress", "other", "s_2"),
  kept_phases = NULL
)
}
\arguments{
\item{kept_phases}{}
}
\description{
Plot CC
}
